updated Test cases, added numbers
