+++ tests/test_el_salvador.py
@@ -9,6 +9,7 @@ def get_driver():
     yield es
 
 
+# T1
 def test_driver_crash():
     try:
         get_driver
@@ -18,6 +19,7 @@ def test_driver_crash():
         assert False
 
 
+# T2
 def test_multiple_queries(get_driver):
     es = get_driver
     try:
@@ -29,30 +31,35 @@ def test_multiple_queries(get_driver):
         assert False
 
 
+# T3
 def test_output_list(get_driver):
     es = get_driver
     data = es.date(day=10, month=10, year=2020)
     assert isinstance(data, list)
 
 
+# T4
 def test_output_value(get_driver):
     es = get_driver
     data = es.date(day=10, month=10, year=2020)
     assert isinstance(data[0]['value'], float)
 
 
+# T5
 def test_output_dict_size(get_driver):
     es = get_driver
     data = es.date(day=10, month=10, year=2020)
     assert 4 == len(data[0])
 
 
+# T6
 def test_output_tz_aware(get_driver):
     es = get_driver
     data = es.date(day=10, month=10, year=2020)
     assert data[0]['ts'].tzinfo is not None
 
 
+# T7
 def test_error_future(get_driver):
     es = get_driver
     try:
@@ -63,6 +70,7 @@ def test_error_future(get_driver):
         assert True
 
 
+# T8
 def test_error_illegal_date(get_driver):
     es = get_driver
     try:
@@ -73,6 +81,7 @@ def test_error_illegal_date(get_driver):
         assert True
 
 
+# T9
 def test_error_backwards_dates(get_driver):
     es = get_driver
     try:
